Moving api Http proxy to its own library
