+++ src/server.js
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom/server';
 import config from './config';
 import favicon from 'serve-favicon';
 import compression from 'compression';
-import httpProxy from 'express-http-proxy';
+import httpProxy from './helpers/api/httpProxy';
 import session from 'express-session';
 import redisStore from 'connect-redis';
 import path from 'path';
@@ -26,46 +26,6 @@ const app = new Express();
 const server = new http.Server(app);
 const RedisStore = redisStore(session);
 
-const VALID_END_POINTS_FOR_SESSION = /\/(login|signups\/complete\/|invitations\/accept)/;
-
-/**
- * Check if API response can set session
- *
- * @param {Object} req
- * @param {Object} res
- * @return {Boolean}
- */
-const canSetSession = (req, res) => {
-  return req.method === 'POST' &&
-        res.statusCode === 200 &&
-        VALID_END_POINTS_FOR_SESSION.test(req.path);
-};
-
-const proxy = httpProxy(`http://${(process.env.HOST || 'localhost')}`, {
-  port: config.apiPort,
-  forwardPath: (req) => {
-    const originalPath = require('url').parse(req.url).path;
-
-    return '/api/v1' + originalPath;
-  },
-  intercept: (rsp, data, req, res, callback) =>{
-    let respondData;
-    if (canSetSession(req, res)) {
-      respondData = JSON.parse(data.toString('utf8'));
-      req.session.user_id = respondData.id;
-      respondData = JSON.stringify(respondData);
-    }
-    callback(null, respondData || data);
-  },
-  decorateRequest: (req) => {
-    if (req.session.user_id) {
-      req.headers['X-katuma-user-id'] = req.session.user_id;
-    }
-    return req;
-  },
-  preserveReqSession: true
-});
-
 app.use(compression());
 app.use(favicon(path.join(__dirname, '..', 'static', 'favicon.ico')));
 
@@ -84,7 +44,7 @@ app.use('/api/v1/logout', (req, res) => {
 });
 
 // Proxy to API server
-app.use('/api/v1', proxy);
+app.use('/api/v1', httpProxy);
 
 app.use((req, res, next) => {
   // Nasty bug related with this issue:
